ai/requirements.txt
numpy
pillow
python-magic
fpdf
essential_generators
requests
pandas
tensorflow
sklearn